refactor(session): extract constants + pure helpers into separate modules
Token optimization: splitta session.js in 3 file per ridurre il
context consumato quando Claude Code legge/edita il session engine.

Nuovo: apps/backend/routes/sessionConstants.js (58 LOC)
- isRoundModelEnabled, GRID_SIZE, DEFAULT_HP/AP/MOD/DC/GUARDIA/INITIATIVE
- DEFAULT_FACING, VALID_FACINGS, JOB_INITIATIVE, JOB_STATS, ASSIST_WINDOW_TURNS

Nuovo: apps/backend/routes/sessionHelpers.js (252 LOC)
- rollD20, clampPosition, normaliseUnit, buildDefaultUnits, normaliseUnitsPayload
- resolveAttack, timestampStamp, publicSessionView
- buildTurnOrder, nextUnitId, manhattanDistance, pickLowestHpEnemy, stepTowards
- isBackstab, facingFromMove

Modificato: apps/backend/routes/session.js (1967 → 1665 LOC, -15%)
- Rimossa definizione inline di tutte le funzioni pure e costanti
- Aggiunti 2 require per sessionConstants + sessionHelpers
- createSessionRouter closure intatto (funzioni di combattimento che
  usano closure-scoped session Map, traitRegistry, rng)

Zero cambiamento funzionale: tutti i 162 test verdi. Le funzioni
esportate dal session.js module.exports restano identiche (le re-esporta
da sessionHelpers).

Impatto token: quando Claude edita session.js (il file piu' letto
durante combat edits), legge 1665 LOC invece di 1967 LOC. Le
costanti e gli helper sono leggibili solo se serve (Read mirato).

Validazione:
- node --test tests/api/sessionRoundEndpoints.test.js + action + turn/end -> 24/24
- node --test tests/ai/*.test.js tests/services/*.test.js -> 138/138
- governance strict -> 0 errors
- prettier -> verde

Rollback: git revert <sha> rimette tutto inline in session.js.
I file sessionConstants.js e sessionHelpers.js sono additive e
non hanno altri consumatori fuori da session.js.
