+++ apps/backend/routes/sessionHelpers.js
@@ -1,0 +1,248 @@
+// Session engine pure helpers — extracted from session.js for token optimization.
+// All functions here are pure (no closure deps, no I/O, no session Map access).
+// They depend only on constants from sessionConstants.js.
+
+'use strict';
+
+const {
+  GRID_SIZE,
+  DEFAULT_HP,
+  DEFAULT_AP,
+  DEFAULT_MOD,
+  DEFAULT_DC,
+  DEFAULT_GUARDIA,
+  DEFAULT_INITIATIVE,
+  DEFAULT_FACING,
+  VALID_FACINGS,
+  JOB_INITIATIVE,
+  JOB_STATS,
+} = require('./sessionConstants');
+
+const { DEFAULT_ATTACK_RANGE } = require('../services/ai/policy');
+
+function rollD20(rng) {
+  return Math.floor(rng() * 20) + 1;
+}
+
+function clampPosition(x, y) {
+  return {
+    x: Math.min(Math.max(0, Number(x) || 0), GRID_SIZE - 1),
+    y: Math.min(Math.max(0, Number(y) || 0), GRID_SIZE - 1),
+  };
+}
+
+function normaliseUnit(raw, fallbackIndex) {
+  const input = raw && typeof raw === 'object' ? raw : {};
+  const id = String(input.id || `unit_${fallbackIndex + 1}`);
+  const position =
+    input.position && typeof input.position === 'object'
+      ? clampPosition(input.position.x, input.position.y)
+      : fallbackIndex === 0
+        ? { x: 0, y: 0 }
+        : { x: GRID_SIZE - 1, y: GRID_SIZE - 1 };
+  const traits = Array.isArray(input.traits) ? input.traits.filter(Boolean).map(String) : [];
+  const ap = Number.isFinite(Number(input.ap)) ? Number(input.ap) : DEFAULT_AP;
+  const job = input.job ? String(input.job) : 'unknown';
+  const jobStats = JOB_STATS[job] || {};
+  const attackRange = Number.isFinite(Number(input.attack_range))
+    ? Number(input.attack_range)
+    : Number.isFinite(Number(jobStats.attack_range))
+      ? Number(jobStats.attack_range)
+      : DEFAULT_ATTACK_RANGE;
+  const hp = Number.isFinite(Number(input.hp)) ? Number(input.hp) : DEFAULT_HP;
+  const maxHp = Number.isFinite(Number(input.max_hp)) ? Number(input.max_hp) : hp;
+  const status =
+    input.status && typeof input.status === 'object'
+      ? { ...input.status }
+      : { panic: 0, rage: 0, stunned: 0, focused: 0, confused: 0, bleeding: 0, fracture: 0 };
+  const initiative = Number.isFinite(Number(input.initiative))
+    ? Number(input.initiative)
+    : Number.isFinite(Number(JOB_INITIATIVE[job]))
+      ? Number(JOB_INITIATIVE[job])
+      : DEFAULT_INITIATIVE;
+  const rawFacing = input.facing ? String(input.facing).toUpperCase() : null;
+  const facing = VALID_FACINGS.has(rawFacing) ? rawFacing : fallbackIndex === 0 ? 'S' : 'N';
+  return {
+    id,
+    species: input.species ? String(input.species) : 'unknown',
+    job,
+    traits,
+    hp,
+    max_hp: maxHp,
+    status,
+    ap,
+    ap_remaining: Number.isFinite(Number(input.ap_remaining)) ? Number(input.ap_remaining) : ap,
+    mod: Number.isFinite(Number(input.mod)) ? Number(input.mod) : DEFAULT_MOD,
+    dc: Number.isFinite(Number(input.dc)) ? Number(input.dc) : DEFAULT_DC,
+    guardia: Number.isFinite(Number(input.guardia)) ? Number(input.guardia) : DEFAULT_GUARDIA,
+    attack_range: attackRange,
+    initiative,
+    facing,
+    position,
+    controlled_by: input.controlled_by ? String(input.controlled_by) : 'player',
+  };
+}
+
+function buildDefaultUnits() {
+  return [
+    normaliseUnit(
+      {
+        id: 'unit_1',
+        species: 'velox',
+        job: 'skirmisher',
+        traits: ['zampe_a_molla'],
+        position: { x: 0, y: 0 },
+        controlled_by: 'player',
+      },
+      0,
+    ),
+    normaliseUnit(
+      {
+        id: 'unit_2',
+        species: 'carapax',
+        job: 'vanguard',
+        traits: ['pelle_elastomera'],
+        position: { x: 5, y: 5 },
+        controlled_by: 'sistema',
+      },
+      1,
+    ),
+  ];
+}
+
+function normaliseUnitsPayload(raw) {
+  if (!Array.isArray(raw) || raw.length === 0) {
+    return buildDefaultUnits();
+  }
+  return raw.map((entry, index) => normaliseUnit(entry, index));
+}
+
+function resolveAttack({ actor, target, rng }) {
+  const die = rollD20(rng);
+  const roll = die + (actor.mod || 0);
+  const dc = target.dc ?? target.dc_difesa ?? 10 + (target.mod || 0);
+  const mos = roll - dc;
+  const hit = mos >= 0;
+  let pt = 0;
+  if (hit) {
+    if (die >= 15 && die <= 19) pt += 1;
+    if (die === 20) pt += 2;
+    pt += Math.floor(mos / 5);
+  }
+  return { die, roll, mos, hit, dc, pt };
+}
+
+function timestampStamp(date) {
+  const iso = date.toISOString();
+  return iso
+    .replace(/[-:]/g, '')
+    .replace(/\..*Z$/, '')
+    .replace('T', '_');
+}
+
+function publicSessionView(session) {
+  return {
+    session_id: session.session_id,
+    turn: session.turn,
+    active_unit: session.active_unit,
+    turn_order: session.turn_order || [],
+    turn_index: session.turn_index || 0,
+    units: session.units,
+    grid: session.grid,
+    grid_size: session.grid.width,
+    log_events_count: session.events.length,
+  };
+}
+
+function buildTurnOrder(units) {
+  return units
+    .map((u, idx) => ({ id: u.id, init: Number(u.initiative) || 0, idx }))
+    .sort((a, b) => b.init - a.init || a.idx - b.idx)
+    .map((e) => e.id);
+}
+
+function nextUnitId(session) {
+  const order = session.turn_order;
+  if (Array.isArray(order) && order.length > 0) {
+    const n = order.length;
+    let idx = Number.isFinite(session.turn_index) ? session.turn_index : -1;
+    for (let i = 0; i < n; i += 1) {
+      idx = (idx + 1) % n;
+      const candidateId = order[idx];
+      const unit = session.units.find((u) => u.id === candidateId);
+      if (unit && unit.hp > 0) {
+        session.turn_index = idx;
+        return candidateId;
+      }
+    }
+    return null;
+  }
+  const units = session.units;
+  if (!units.length) return null;
+  const currentIdx = units.findIndex((u) => u.id === session.active_unit);
+  const nextIdx = currentIdx < 0 ? 0 : (currentIdx + 1) % units.length;
+  return units[nextIdx].id;
+}
+
+function manhattanDistance(a, b) {
+  return Math.abs(a.x - b.x) + Math.abs(a.y - b.y);
+}
+
+function pickLowestHpEnemy(session, actor) {
+  const enemies = session.units.filter((u) => u.id !== actor.id && u.hp > 0);
+  if (!enemies.length) return null;
+  return enemies.reduce((lowest, candidate) => {
+    if (!lowest) return candidate;
+    return candidate.hp < lowest.hp ? candidate : lowest;
+  }, null);
+}
+
+function stepTowards(from, to) {
+  const next = { ...from };
+  if (from.x !== to.x) {
+    next.x += from.x < to.x ? 1 : -1;
+  } else if (from.y !== to.y) {
+    next.y += from.y < to.y ? 1 : -1;
+  }
+  return clampPosition(next.x, next.y);
+}
+
+function isBackstab(actor, target) {
+  if (!actor?.position || !target?.position) return false;
+  const f = target.facing || DEFAULT_FACING;
+  const dx = actor.position.x - target.position.x;
+  const dy = actor.position.y - target.position.y;
+  if (f === 'N') return dy > 0;
+  if (f === 'S') return dy < 0;
+  if (f === 'E') return dx < 0;
+  if (f === 'W') return dx > 0;
+  return false;
+}
+
+function facingFromMove(from, to) {
+  const dx = to.x - from.x;
+  const dy = to.y - from.y;
+  if (dx === 0 && dy === 0) return null;
+  if (Math.abs(dx) >= Math.abs(dy)) {
+    return dx > 0 ? 'E' : 'W';
+  }
+  return dy > 0 ? 'S' : 'N';
+}
+
+module.exports = {
+  rollD20,
+  clampPosition,
+  normaliseUnit,
+  buildDefaultUnits,
+  normaliseUnitsPayload,
+  resolveAttack,
+  timestampStamp,
+  publicSessionView,
+  buildTurnOrder,
+  nextUnitId,
+  manhattanDistance,
+  pickLowestHpEnemy,
+  stepTowards,
+  isBackstab,
+  facingFromMove,
+};
