+++ reports/docs/governance_drift_report.json
@@ -1,5 +1,5 @@
 {
-  "generated_at": "2026-04-16T00:32:14+00:00",
+  "generated_at": "2026-04-16T00:41:01+00:00",
   "summary": {
     "total": 0,
     "errors": 0,
